Notes and Stuff on Relays update to README.md
Likely to go unnoticed, a change to aid branding that is consistent with the long standing descriptor of the purpose and makeup of the protocol. If this change is noticed, it's likely to be appreciated, because it can help more clearly and succinctly convey what the protocol is. The abbreviation still works NOSTR is made up of the following capitals: "NOtes and STuff on Relays". Still in sequence and with appropriate emphasis. So we still call it Nostr. We still know it's abbreviated. We still know it's a protocol with notes and stuff and relays. And we still know it's what we've always used. It's just now easier to tie to a tagline that is just as descriptive without such a long sentence. This will be helpful for newly onboarded users and for when referenced in content and media.